lecture 129

+++ backend/routes/posts.js
@@ -12,6 +12,7 @@ const MIME_TYPE_MAP = {
   'image/jpg': 'jpg'
 }
 
+// Image Storage Configurations
 const storage = multer.diskStorage({
   destination: (req, file, cb) => {
     const isValid = MIME_TYPE_MAP[file.mimetype];
@@ -28,6 +29,7 @@ const storage = multer.diskStorage({
   }
 });
 
+// Create Post Method
 router.post("",
   checkAuth,
   multer({storage: storage}).single("image"),
@@ -36,7 +38,8 @@ router.post("",
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/uploads/images/' + req.file.filename
+    imagePath: url + '/uploads/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(result => {
     console.log(result);
@@ -47,9 +50,15 @@ router.post("",
         id: result._id,
       }
     });
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Please try again later."
+    })
   });
 });
 
+// Fetch Posts Method
 router.get('', (req,res,next) => {
   const pageSize = +req.query.pageSize;
   const page = +req.query.page;
@@ -71,20 +80,38 @@ router.get('', (req,res,next) => {
       posts: fetchedPosts,
       maxPosts: count
     });
-  });
-});
-
-router.delete('/:id', checkAuth, (req,res,next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
+  })
+  .catch(error => {
     res.status(200).json({
-      message: "Post Deleted"
+      message: "Please try again later."
     });
   });
-
 });
 
+// Post Delete Method
+router.delete('/:id', checkAuth, (req,res,next) => {
+  Post.deleteOne({_id: req.params.id, creator: req.userData._id}).then(result => {
+    console.log(result);
+    if (result.n > 0) {
+      res.status(200).json({
+        message: "Deletion Successful"
+      });
+    } else {
+      res.status(401).json({
+        message: "Deletion Failed."
+      });
+    }
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Please try again later."
+    })
+  });
+});
+
+// Post Update Method
 router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
+
   let imagePath = req.body.imagePath;
   if (req.file) {
     const url = req.protocol + "://" + req.get("host");
@@ -96,13 +123,25 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
     content: req.body.content,
     imagePath: imagePath
   })
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({
-      message: "Update Successful"
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({
+        message: "Update Successful"
+      });
+    } else {
+      res.status(401).json({
+        message: "Update Failed."
+      });
+    }
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Please try again later."
     })
   });
 })
 
+// Fetch Individual Post Method
 router.get("/:id", (req, res, next) => {
   Post.findById(req.params.id).then(post => {
     if(post) {
@@ -110,6 +149,11 @@ router.get("/:id", (req, res, next) => {
     } else {
       res.status(404).json({message: 'Post not found'});
     }
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: 'Please try again later.'
+    });
   });
 })
